% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the number of accidents from FARS data for a given year and state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number with the state.num code for the state to be mapped}

\item{year}{a number for the year for FARS data accidents to be plotted in the map.}
}
\value{
This function returns a map in the Plots tab using data from the FARS data
   for the \code{year} input, filtered for the STATE column = \code{year}state.num
   If there is no file for an element of \code{years} then a warning is given for
   that year.
}
\description{
This function takes input corresponding to a state number and year rows and
returns a map of the FARS data with dots at the longitude an lattitude for each
fatal accident from that year.
}
\examples{
\dontrun{fars_map_state(55,2013)}

}
