% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the .csv file with FARS data}
}
\value{
This function returns a tibble with the FARS data
}
\description{
This function reads in a .csv file containing US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System ("FARS") data and returns a
tibble (or error message if file does not exist).
}
\examples{
\dontrun{fars_read("accident_2014.csv")}

}
