% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in the month & year for each accident from a FARS data file for multiple years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector containing the years for FARS data to be included in the output}
}
\value{
This function returns a list of tibbles containing the month and year columns
   from the FARS data file for that year.   If there is no FARS data file for an
   element of \code{years} then NULL is returned for list item corresponding to that
   element.
}
\description{
This function converts a numeric vector to a list of tibbles containing the month and
year columns from the FARS data file for that year.
}
\examples{
\dontrun{fars_read_years(c(2013,2015))}

}
