% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents by month by year for given years of FARS data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector containing the years for FARS data to be included in the output}
}
\value{
This function returns a tibble with 12 rows with a MONTH column and columns
   for each year in \code{years} for which there is a FARS data file for that year.
   If there is no file for an element of \code{years} then a warning is given for
   that year.
}
\description{
This function converts a numeric vector to a table containing months as rows and
years as columns which provides the number of accidents from FARS data.
}
\examples{
\dontrun{fars_read_years(c(2013,2015))}

}
