% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name for FARS data file given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year for the FARS file}
}
\value{
This function returns a character string a string of the form "accident
   year_[YEAR].csv.bz2"
}
\description{
This function converts a year to a string of the form "accident year_[YEAR].csv.bz2"
which corresponds to the naming convention for FARS files.
}
\examples{
\dontrun{make_filename(2013)}

}
